Enable libva on Windows
I was not able to get the driver to work at all.
It just crashes a lot, but since it has no averse side-effects
(hopefully), let's just enable it by default.

+++ scripts.d/50-vaapi/99-finalize.sh
@@ -3,7 +3,6 @@
 SCRIPT_SKIP="1"
 
 ffbuild_enabled() {
-    [[ $TARGET != linux* ]] && return -1
     return 0
 }
 
@@ -12,6 +11,8 @@ ffbuild_dockerdl() {
 }
 
 ffbuild_dockerbuild() {
-    rm "$FFBUILD_PREFIX"/lib/lib*.so* || true
-    rm "$FFBUILD_PREFIX"/lib/*.la || true
+    if [[ $TARGET == linux* ]]; then
+        rm "$FFBUILD_PREFIX"/lib/lib*.so* || true
+        rm "$FFBUILD_PREFIX"/lib/*.la || true
+    fi
 }
